Rename extract templates to template-info
