+++ lib/linter.js
@@ -3,7 +3,7 @@ import * as contentTagUtils from 'content-tag-utils';
 
 import ModuleStatusCache from './-private/module-status-cache.js';
 import TodoHandler from './-private/todo-handler.js';
-import { templateInfoForTemplateFile, templateInfoForScript } from './extract-templates.js';
+import { templateInfoForTemplateFile, templateInfoForScript } from './template-info.js';
 import { isHBS, isHTML, isGlimmerFileExtension, isScript } from './file-types.js';
 import PrettyFormatter from './formatters/pretty.js';
 import { getProjectConfig, getRuleFromString } from './get-config.js';
