+++ lib/template-info.js
